+++ lib/experimental/editor-settings.php
@@ -38,6 +38,9 @@ function gutenberg_enable_experiments() {
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-dashboard-widgets' ) ) {
 		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalDashboardWidgets = true', 'before' );
 	}
+	if ( gutenberg_is_experiment_enabled( 'gutenberg-suggestion-mode' ) ) {
+		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalSuggestionMode = true', 'before' );
+	}
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-real-time-collaboration' ) ) {
 		$collaboration_enabled = ! (
 			'site-editor.php' === $pagenow ||